Merge in RFRM/pay from MGMPROCE-1038-1042 to develop

* commit '4308abec505401e85fb13754bd9461ebfb5db6fe':
  Правка ошибки с записью статусов в БД - пишет 2 раза в history, и начальное заполнение processed(null) in CorePayKafkaModel
  Правка ошибки с записью статусов в БД - пишет 2 раза в history, и начальное заполнение processed(null) in CorePayKafkaModel

+++ src/test/java/ru/vtb/msa/rfrm/integration/rfrmkafka/mapper/QuestionnairesMapperImplTest.java
@@ -45,7 +45,7 @@ class QuestionnairesMapperImplTest {
 
         return CorePayKafkaModel
                 .builder()
-                .rewardId(random.nextInt())
+                .rewardId(random.nextLong())
                 .mdmId(random.nextLong())
                 .questionnaireId(UUID.randomUUID())
                 .recipientTypeId(3)
